GalerinhaAPI/requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14.0

# Database
psycopg2-binary>=2.9.9

# Authentication
djangorestframework-simplejwt>=5.3.0

# CORS
django-cors-headers>=4.3.0

# Environment Variables
python-dotenv>=1.0.0

# Utils
django-filter>=23.5

# Dev & Testing
pytest>=7.4.0
pytest-django>=4.5.0
factory-boy>=3.3.0
